covid19api/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
requests
djangorestframework
django-health-check
django-rest-swagger
drf-yasg